Set default values for Socket's `ipvo` and `timeout`
Commit
https://github.com/jtesta/ssh-audit/commit/f44663bfc4a7c70f9a61a8304fd6f16d945403d8
introduced two new arguments to the Socket class, but did not update
the tests, which still relied on the socket class to only require two arguments.

While for `ipvo`the default of `None` is obvious, as in `__init__` it is
checked for it, for `timeout` it was not that obvious.

Luckily, in the README a default of 5 (seconds) is mentioned.

modified:   ssh-audit.py